Modified a lot

+++ src/foundation/C_Crypta.java
@@ -7,19 +7,18 @@ public class C_Crypta {
 	public static void main(String[] args) throws Exception {
 		byte[] ct = CryptoTools.fileToBytes("data/MSG2.ct.txt");
 
-		int[] freq = CryptoTools.getFrequencies(ct);
-		int max_freq = 0;
+		double[] freq = CryptoTools.getFrequencies(ct);
+		double max_freq = 0;
 		char max_char = 0;
 		for (int i = 0; i < freq.length; i++) {
-			System.out.printf("%c ---> frequceny: %d\n", i + 'A', freq[i]);
+			System.out.printf("%c ---> frequceny: %f\n", i + 'A', freq[i]);
 			if (freq[i] > max_freq) {
 				max_freq = freq[i];
 				max_char = (char) ('A' + i);
 			}
 		}
-		
 		System.out.println("=====================================================");
-		System.out.printf("MAX: %c ---> frequceny: %d\n", max_char, max_freq);
+		System.out.printf("MAX: %c ---> frequceny: %f\n", max_char, max_freq);
 		int key = max_char - 'E';
 		if (key < 0) key += 26;
 		System.out.println("key = " + key);
